requirements.txt
# OptikR Requirements
# Install: pip install -r requirements.txt
# Python 3.10+ compatible

# ============================================================================
# GUI Framework
# ============================================================================
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0

# ============================================================================
# Deep Learning & AI
# ============================================================================
# NOTE: PyTorch is installed separately based on your hardware
# For manual installation:
# - CPU only: pip install -r requirements-cpu.txt
# - GPU (CUDA 12.4): pip install -r requirements-gpu.txt

# ============================================================================
# OCR Engines
# ============================================================================
easyocr>=1.7.0
paddleocr>=2.7.0
manga-ocr>=0.1.11
pytesseract>=0.3.10
winocr>=0.0.6  # Optional: Windows OCR (wraps Windows.Media.Ocr / DirectML)
rapidocr-onnxruntime>=1.3.0  # Optional: Lightweight PaddleOCR via ONNX Runtime
python-doctr[torch]>=0.8.0  # Optional: Transformer-based OCR (Mindee DocTR)
surya-ocr>=0.4.0,<=0.13.1  # Optional: Vision-transformer OCR (90+ languages) — pinned: 0.14+ requires torch>=2.7 (no CUDA 12.4 wheels)
# mokuro — installed automatically by the plugin system with --no-deps
# to avoid overwriting CUDA torch.  Manual install:
#   pip install mokuro --no-deps
#   pip install fire loguru natsort pyclipper shapely torchsummary yattag
# mokuro>=0.2.1

# ============================================================================
# Vision Mode (Qwen3-VL)
# ============================================================================
# qwen-vl-utils — installed with --no-deps to avoid pulling in packages
# that could overwrite CUDA torch.  Manual install:
#   pip install qwen-vl-utils --no-deps
#   pip install av
# qwen-vl-utils>=0.0.14

# ============================================================================
# Translation
# ============================================================================
transformers>=4.57.0,<5.0.0
sentencepiece>=0.1.99  # Required by transformers/MarianMT
huggingface-hub>=0.23.0  # Required by transformers
# accelerate — installed automatically by the plugin system with --no-deps
# to avoid overwriting CUDA torch.  Manual install:
#   pip install accelerate --no-deps
#   pip install pyyaml safetensors
# accelerate>=0.26.0

# ============================================================================
# Image Processing
# ============================================================================
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# ============================================================================
# Screen Capture
# ============================================================================
bettercam>=1.0.0  # GPU-accelerated capture (AMD/NVIDIA, Desktop Duplication API)
mss>=9.0.1

# ============================================================================
# Package Management
# ============================================================================
packaging>=21.0

# ============================================================================
# System Utilities
# ============================================================================
psutil>=5.9.0
pystray>=0.19.0
py-cpuinfo>=9.0.0

# ============================================================================
# Networking
# ============================================================================
requests>=2.31.0
urllib3>=2.0.0  # Required by requests

# ============================================================================
# Security
# ============================================================================
cryptography>=42.0.0

# ============================================================================
# Windows-specific
# ============================================================================
pywin32>=306

# ============================================================================
# Translation Services
# ============================================================================
deep-translator>=1.11.0
deepl>=1.16.0
google-cloud-translate>=3.12.0  # Optional: Google Cloud Translation API (premium)

# ============================================================================
# Text Processing
# ============================================================================
textdistance>=4.5.0
pyspellchecker>=0.7.2  # Optional: Used by spell checker plugin

# ============================================================================
# Performance Monitoring
# ============================================================================
nvidia-ml-py>=12.0.0  # Optional: NVIDIA GPU monitoring

# ============================================================================
# GPU Acceleration (Optional - Uncomment if needed)
# ============================================================================
# cupy>=12.0.0  # CUDA-accelerated NumPy
# pyopencl>=2023.1  # OpenCL support
# numba>=0.57.0  # JIT compiler

# ============================================================================
# Audio Translation Plugin (Optional - Uncomment if needed)
# ============================================================================
# openai-whisper>=20231117
# pyaudio>=0.2.13
# pyttsx3>=2.90
# webrtcvad>=2.0.10  # Voice Activity Detection
# TTS>=0.22.0  # Better TTS quality
